module-folder was moved
