+++ ydb/library/yql/providers/yt/provider/ya.make
@@ -50,7 +50,7 @@ PEERDIR(
     yt/cpp/mapreduce/interface
     ydb/library/yql/ast
     ydb/library/yql/core/extract_predicate
-    ydb/library/yql/module
+    ydb/library/yql/udfs/common/module
     ydb/library/yql/public/udf
     ydb/library/yql/public/udf/tz
     ydb/library/yql/sql
