+++ ydb/library/yql/udfs/common/ya.make
@@ -11,6 +11,7 @@ RECURSE(
     json
     json2
     math
+    module
     knn
     pire
     protobuf
