+++ ydb/library/yql/ya.make
@@ -3,7 +3,6 @@ RECURSE(
     core
     dq
     minikql
-    module
     parser
     protos
     providers
